+++ src/sc_crawler/vendors/hcloud.py
@@ -63,7 +63,7 @@ def _server_cpu(server_name):
     if server_name.upper() in ["CPX12", "CPX22", "CPX32", "CPX42", "CPX52", "CPX62"]:
         return ("AMD", None, None)
     if server_name.upper() in ["CAX11", "CAX21", "CAX31", "CAX41"]:
-        return ("AMD", "Ampere Altra", None)
+        return ("Ampere", "Ampere Altra", None)
     if server_name.upper() in ["CCX13", "CCX23", "CCX33", "CCX43", "CCX53", "CCX63"]:
         return ("AMD", None, None)
     raise ValueError("Unknown Hetzner Cloud server name: " + server_name)